+++ include/geos/operation/relate/EdgeEndBundle.h
@@ -93,11 +93,11 @@ protected:
      *  - otherwise, the attribute is NULL.
      *
      */
-    void computeLabelOn(int geomIndex,
+    void computeLabelOn(uint8_t geomIndex,
                         const algorithm::BoundaryNodeRule& boundaryNodeRule);
 
-    void computeLabelSides(int geomIndex);
-    void computeLabelSide(int geomIndex, int side);
+    void computeLabelSides(uint8_t geomIndex);
+    void computeLabelSide(uint8_t geomIndex, uint32_t side);
 };
 
 } // namespace geos:operation:relate
