+++ src/noding/NodingIntersectionFinder.cpp
@@ -79,7 +79,7 @@ NodingIntersectionFinder::processIntersections(
     /**
      * Check for an intersection between two vertices which are not both endpoints.
      */
-    long long segDiff = segIndex1 - segIndex0;
+    long long segDiff = static_cast<long long>(segIndex1 - segIndex0);
     bool isAdjacentSegment = isSameSegString && std::abs(segDiff) <= 1;
     bool isInteriorVertexInt = (!isAdjacentSegment) &&
                                isInteriorVertexIntersection(p00, p01, p10, p11, isEnd00, isEnd01, isEnd10, isEnd11);
